3 of the projects I worked on in school
the jbauer5-se3103-project4-master file is a java program that I built for my final project of the a class where we had to design a basic space invaders game and give it some unique features, I added additional projectiles for the enemies as well as obstacles to protect the player. Also once a player reaches a score threshold of 100 the enemies move faster but the player gets more guns. 

The ContactManager file is a .net core application using the MVC format and creates a local host web application that is connected to a local database and can be updated with new contacts.

the Project8 file is a windows form application that is used to create a list of current employees and managers and display them in a list box.